adk_perfis_ia/pyproject.toml
[project]
name = "adk-perfis-ia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.1.1",
    "google-genai>=1.17.0",
    "python-dotenv>=1.1.0",
]
